+++ PetFamily.Backend/FileService/src/FileService/Features/CompleteMultipartUpload.cs
@@ -1,8 +1,6 @@
-using Amazon.S3;
-using Amazon.S3.Model;
 using FileService.API;
-using FileService.Core;
-using FileService.Endpoints;
+using FileService.API.Endpoints;
+using FileService.Core.Models;
 using FileService.Infrastructure.Providers;
 using FileService.Infrastructure.Repositories;
 using FileService.Jobs;
