+++ PetFamily.Backend/FileService/src/FileService/Features/UploadPresignedUrl.cs
@@ -1,10 +1,8 @@
 using FileService.API;
-using FileService.Core;
-using FileService.Endpoints;
+using FileService.API.Endpoints;
+using FileService.Core.Models;
 using FileService.Infrastructure.Providers;
 using FileService.Infrastructure.Repositories;
-using FileService.Jobs;
-using Hangfire;
 
 namespace FileService.Features;
