+++ PetFamily.Backend/FileService/src/FileService/Infrastructure/Repositories/FileRepository.cs
@@ -1,6 +1,7 @@
 using CSharpFunctionalExtensions;
 using FileService.Core;
 using FileService.Core.CustomErrors;
+using FileService.Core.Models;
 using FileService.Infrastructure.MongoDataAccess;
 using MongoDB.Driver;
